Updating validation errors and logging (#3965)
## Summary
Fixes #3867 

### Time to review: __5 mins__

## Changes proposed
The validation code currently logs errors for items that fail. We don't
want to log errors for those items as they will condition developers to
ignore errors that show up in the logs. This change is to update the
logging to make it more specific as to what the problem is and modify
the logging level.

## Context for reviewers
> Running the pipeline as usual should have validation errors that are
no longer logged as an error

## Additional information
> <img width="1477" alt="image"
src="https://github.com/user-attachments/assets/41c711d3-0d1f-4607-87bd-734509fe65a2"
/>