[Issue #3994] Fix agency data during transforms (#3999)
## Summary
Fixes #3994

### Time to review: __5 mins__

## Changes proposed
Modified the way we determine an agency is a test agency to be based on
a prefix rather than exact match (eg. if we configure `ABC`, `ABC-123`
would now be a test agency as well)

Modified the transform hierachy task to be able to unset top level
agency if it doesn't match anything (fixes some weird data issues we're
seeing)

Modified the transform hierachy to also update `is_test_agency` so that
everything always gets reprocessed each run

## Context for reviewers
The transform agency logic happens in two passes, first we look at all
of the agency changes from the `TGROUPS` table and apply updates /
create new agencies.

We then have a `TransformAgencyHierarchy` task that runs to do some
post-processing on all agencies. I mostly put these fixes in the latter
bit so that any of the weird issues we've noticed get cleaned up.

---------

Co-authored-by: freddieyebra <freddieyebra@navapbc.com>