[Issue #4004] Fix transform problem with simulatenous delete + insert (#4024)
## Summary
Fixes #4004

### Time to review: __15 mins__

## Changes proposed
Modify the transformations for the applicant type, funding category, and
funding instrument transformations to process all deletes separate from
inserts/updates

## Context for reviewers
This fixes a bug where if we received both a delete + insert for one of
these tables at the same time, we'd delete, but be unable to insert.
This happens anytime an opportunity summary is updated in the old system
due to how it handles versioning.

For a much more in-depth description of the bug, see
https://github.com/HHS/simpler-grants-gov/issues/4004#issuecomment-2682991160

By separating the deletes and inserts, we don't hit the weird
delete-not-yet-processed caching issue.