[Issue #4021] Create a task that validates agency data (#4077)
## Summary
Fixes #4021

### Time to review: __5 mins__

## Changes proposed
Created a task that validates agency data for a few known scenarios
we've encountered

## Context for reviewers
This work is a follow-up to a discussion we had about oddities we've
seen with agency data, and mainly aims for us to better be aware of data
ending up in a weird state. While we've fixed some issues with the data,
and fixed processes that build it, we know there are a few issues still
present somewhere in the logic, and want to use this as a mechanism to
help detect it.