[Issue #3936] Product roadmap page (#4020)
## Summary
Fixes #3936

### Time to review: __25 mins__

## Changes proposed
This PR implements a new product roadmap page at /roadmap based on this
design here
https://www.figma.com/design/FGxWtAgToKhehLJCiuy1zL/Simpler.Grants.gov-Project?node-id=3422-31812.

## Context for reviewers
This PR only implements the new page at /roadmap and only implements the
ability to land on this page by direct URL. Follow up work is in place
in #4091 to implement replacing the existing /process page with the
roadmap page.

To verify this change, run the frontend locally and navigate to
localhost:3000/roadmap and ensure the content renders as expected.

## Additional information
![Screen Recording 2025-03-10 at 10 40
43 AM](https://github.com/user-attachments/assets/cafcd608-bf07-47da-8abc-a1aa65374421)