Let Flask pick HTTPS correctly based on if the service is configured … (#4286)
## Summary
Fixes #4263 

### Time to review: __5 mins__

## Changes proposed
Our Login.gov integration has to craft a redirect URL that Login.gov
will send the user back to after login. This allows for control of the
scheme for that URL to account for environments where we have HTTPS in
front of the API vs not.

## Context for reviewers
If you set `LOGIN_GOV_REDIRECT_SCHEME=https` in `api/override.env` and
run through the Login process you'll get redirected to a failed
HTTPS://localhost:3000 URL