[Unticketed] Remove dependency on pytest-watch (#4295)
## Summary

### Time to review: __3 mins__

## Changes proposed
Removed pytest-watch which hadn't been updated since 2018 and was
causing issues building due to a recent change in setuptools

## Context for reviewers
This just removes the ability to watch tests and have it constantly run
tests everytime your code changes. If we really want that functionality,
we can look into https://pypi.org/project/pytest-watcher/ which seems to
be the spiritual successor to pytest-watch and is still maintained.