[Issue #4168] Update competition/form/application factories to better support test setup (#4290)
## Summary
Fixes #4168

### Time to review: __5 mins__

## Changes proposed
* Form factory now has a valid json schema + UI schema
* Adjusted schema of form response as UI schema is a list, not a dict
(make it raw type to be JSON)
* Adjusted factories ordering to be a bit better organized
* Cleaned up a few tests that were doing a lot more than necessary to
setup data

## Context for reviewers
Note that the rearranging of factories is a sorta-precursor to work I'm
thinking we need to do to split this file up a bit (at least move the
staging/legacy table factories to a separate file).

The test updates are mostly to remove redundant factory usage. For
example, in some tests we were creating records, then passing those to
another factory, and then passing that to a 3rd factory. Unless there is
a specific setup on one of those earlier records, we don't need to do
this as the factories themselves create all of their dependencies.
Before this change a few of those dependencies were missing, so the
tests weren't wrong, just didn't fully have the support from these
changes.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>