[unticketed] Update progress link on homepage. (#4291)
## Summary
no ticket but related to #4091 

### Time to review: __5 mins__

## Changes proposed
Fix a link on the home page that still tries to redirect to the /process
and redirect it to /roadmap instead.

## Context for reviewers
There was an instance of replacing all of the links that go to /process
page to /roadmap. To test this:
- Checkout this branch `git checkout fyebra/fix-roadmap-link`
- Run the frontend locally with `npm run dev`
- Open the homepage and click the "learn about our progress" button.
- Ensure it redirects to /roadmap page.

## Additional information
> Screenshots, GIF demos, code examples or output to help show the
changes working as expected.