[Issue #4246] Create mock SOAP service for local development (#4278)
## Summary
Fixes #4246

### Time to review: __15 mins__

## Changes proposed
Add a local docker container for running a mock server locally that will
eventually be consumed by the api for local development.

## Context for reviewers
We need a mock server in order to test proxy functionality to grants.gov
SOAP API.

To test this PR follow the instructions in the "Running the mock SOAP
services" section of the README in this PR and ensure the instructions
work as expected.

## Additional information
> Screenshots, GIF demos, code examples or output to help show the
changes working as expected.