[Issue #4317] Remove agencies with no open or forecasted opportunities from options list (#4321)
## Summary
Fixes #{[4317](https://github.com/HHS/simpler-grants-gov/issues/4317)}

### Time to review: __5 mins__

## Changes proposed
The `get_agencies` endpoint has been updated to accept a filter
parameter: `active=true`. When this filter is applied, it will return a
list of sub-agencies and respective top-level agencies that are
associated with at least one opportunity in either the `forecasted` or
`posted` status. If the filter is not provided, the endpoint will return
all agencies.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>