[Issue #4355] Adjust logging in LoadOracleDataTask to use less metrics (#4383)
## Summary
Fixes #4355

### Time to review: __5 mins__

## Changes proposed
Updated logging in LoadOracleDataTask to not put it all in metrics / log
everything at once

## Context for reviewers
New Relic has a limit of ~255 key/value pairs in a log statement. Right
now we're putting about 200+ values into the metric object just for this
one task - which is causing it to break when combined with the ~50 or so
we always add. To work around this, instead of adding all these values
as we go, just log a bunch of them as we go. Only keep metrics for the
overall counts, and just log insert/update/delete metrics when we
calculate them. Still have all the same info, just spread out a bit.