[Issue #4172] Add JSON schema validation to the PUT /applications/:appId/forms/:formId (#4314)
## Summary
Fixes #4172

### Time to review: __5 mins__

## Changes proposed
Adds JSON schema validation utilities

Adds JSON schema validation to the user response for a given application

## Context for reviewers
Note that the validation doesn't prevent or in any way block a users
answers from being filled out, it's just the current set of open issues.
This is so if a user partially fills out a form, we'll note down what is
still needed/incorrect, but not block them from saving their answers. It
also allows the frontend to build out its structure in whatever way it
wants (eg. a user could fill out a form across multiple pages).

When we eventually have a concept of submitting a form, at that point we
would error if there are any outstanding issues.