[Unticketed] Add additional logging context to all logs (#4395)
## Summary

### Time to review: __3 mins__

## Changes proposed
Adds `app_name` and `run_mode` to API logs

## Context for reviewers
Useful for grouping/separating logs in New Relic. The app name helps us
separate logs from the frontend/analytics, and the run mode helps us
distinguish when we're running as an API service versus just a python
script.

## Additional information
Some logs when running the API:
![Screenshot 2025-04-01 at 10 44
21 AM](https://github.com/user-attachments/assets/ac39afce-56fd-44b2-ba81-6bfa41601b19)

Some logs when running as a script:
![Screenshot 2025-04-01 at 10 44
39 AM](https://github.com/user-attachments/assets/a1979303-abc7-4194-8362-eab1b52d885d)