[Issue #4264] Create util / tests for dat to JSONSchema conversion (#4350)
## Summary
Fixes #4264

### Time to review: 30 mins

## Changes proposed
- Add CSV parser to JSONSchema
- Add util to convert XLS to CSV
- Parse these fields based on metadata from DAT entries.
- Add pandas to project for XLS parsing

## Context for reviewers
We parse multiple DAT files to test:
https://grants.gov/forms/form-items-description/fid/668
https://grants.gov/forms/form-items-description/fid/723

We either skip or set unknown data types to `string` (skipped types are
defined in an array in code)

## Additional information
poetry.lock - how should this be updated?
This will eventually be rolled into a CLI tool, and will change
frequently over time. Not by any means a final implementation.