[Issue #4452] Further attempts to prevent saved items caching  (#4498)
## Summary
Fixes #4452

### Time to review: __5 mins__

## Changes proposed
Use NextJS config to prevent pages we want to always re-render from
caching. You don't see the effect of this in npm run dev have to build &
start.

Context for reviewers
This is really tough to test, but the Cache-Control headers are better
behaved under build & start following this change so 🤞🏻