[Issue #4546] Modify search logic to avoid querying for more than 10k records (#4553)
## Summary
Fixes #4546

### Time to review: __10 mins__

## Changes proposed
Modified the logic around parsing the search response to cap the total
results at 10k

Modified the logic around creating a request to the search index to
prevent a user from asking for anything past 10k

## Context for reviewers
If you try to query a search index for any record past 10k, it will
error.

For example, if we tried to do this following query against the search
index, it would error saying you tried to ask for record 10005, which is
more than 10k. There is no way around this limit in a way that still
supports pagination (there is a scroll API, but you have to iterate one
page at a time).

```
GET opportunity-index-alias/_search
{
  "size": 10,
  "from": 9995
}
```

To work around this issue, we changed two things:
1. Even if the search index says it has more than 10k results, cap it to
10k (note our metrics for facet counts are not affected by this and can
still be more than 10k). This way any user like our frontend that
respects the page counts/total records won't hit any issues.
2. If a user does request a page that would go over 10k results then
adjust that page's size so it only goes up to 10k, but not past it (in
the above example, we'd change the size to `5` and it would work). If
they fully request something past 10k, then we'll do a page size of 0,
and start from 10k. This keeps a valid request and means things like
facet counts still work.

## Additional information
In addition to the unit tests, I loaded more than 10k records into my
local search index and queried it with a bunch of things around 10k and
couldn't get any errors to happen.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>