[no ticket] Fix permanent CDN diffs (#4588)
## Current Behavior

The `aws_cloudfront_distribution` resource has a terraform diff on it
every single time we deploy. It's confusing and wastes time. Also it's
unclear if the changes are even applying... since we try to apply them
on every single deploy.

## Desired Behavior

There's only a terraform diff on the resource when you actually change
it!!!

## Changes

I change every attribute that had a permanent diff to the value of the
attribute that terraform keeps changing it to. So that there's no more
diff when you deploy the CDN resource back to back. I'm slightly worried
that these changes make it so that caching is disabled by default, but
we should be setting cache headers on all our objects, to ensure that
the CDN is actually being utilized.

## Testing

These routes are still accessible:

- https://d20lnyycqwjw31.cloudfront.net/
- https://d3e8vod9uic4zv.cloudfront.net/bg.svg