[Issue #3915] add "any" filter option (#4667)
Adds "any" check boxes to all filter accordions except for agency. This checkbox effectively does the same thing as the "clear all" link, removing any currently selected options for a given filter. The box should be checked automatically if no other options are checked for a given filter, and will be unchecked as soon as any option for the filter is selected