[Issue #4243] Allow user to download all attachments as zip (#4485)
## Summary
Fixes #4243

### Time to review: __15 mins__

## Changes proposed
Add a Button to the Downloads section if there are any attachments. That
button sends the user to the zip download URL for that Opportunity.

## Context for reviewers
Doesn't fully work locally because the attachment URLs are not behind
the CDN. But did work with mock attachment data taken from a Dev API
response. To test locally you can point your laptop at the Dev API and
use http://localhost:3000/opportunity/40009

---------

Co-authored-by: Doug Schrashun <dougschrashun@navapbc.com>