[Issue #4479] Add competition transformer (#4694)
## Summary
Fixes #4479

### Time to review: 45 mins

## Changes proposed
Add competition transformer
Create relevant tests for various transformer scenarios
Update factories and conftest


## Context for reviewers
Important note is how a competition connects to an opportunity in the
old system. For various reasons in the ticket, we join the
opportunity_cfda record so we know the opportunity ID. Otherwise the
general transformer pattern is implemented here.

## Additional information
See unit tests for scenarios. Can be broken into smaller PRs if review
proves difficult.