[unticketed] Specify text/plain content type for well known endpoint. (#4834)
## Summary
Fixes unticketed

### Time to review: __5 mins__

## Changes proposed
This change updates the response type for the well-known domain
verification endpoint. Previously the response type was text/html and
this change specifies text/plain.

## Context for reviewers
Currently, the returned text data required markup to for line breaks but
can now support including the \n character in the response.

To verify this pull request:
1. Run the api service locally
2. Run `curl 'http://0.0.0.0:8080/.well-known/pki-validation/dv.txt'`
and ensure the following is returned.
```
data
on
newlines
```
3. Run `curl 'http://0.0.0.0:8080/.well-known/pki-validation/d.txt'` and
verify the following is returned
```
Could not find d.txt
```

## Additional information
> Screenshots, GIF demos, code examples or output to help show the
changes working as expected.