[Issue #4663] Builds an entire custom fluentbit container  (#4792)
## Summary

Relates to #4663

## Motivation

This is a well known way to get logs pushing to multiple locations! In
particular we need logs in New Relic for observability, and Cloudwatch
for auditing.

## Testing

I need whoever approves this to actually run through the testing
directions! I don't want to be the only person who can run this.