[Issue #4919] Add shared modules terraform to changes that will deploy services (#4920)
## Summary

Fixes #4919

## Changes proposed

We have shared module code in terraform that may/can affect any of our
services. But we don't redeploy those services automatically when
changes are made there. This rectifies that and will trigger deploys to
all Platform services when a change is made to any of the shared
modules.