[Issue #4688] Multiauth support for our endpoints (#4954)
## Summary

Fixes #4688

## Changes proposed
* Added a utility for supporting multiple distinct authentication
methods on a single endpoint

## Context for reviewers
Some of our upcoming work will want a way to authenticate as either a
user or with an API key. This started as a simple test to see if that
was possible, and by the time I proved I could do it, most of it was
written, so rather than have follow-up tickets to implement it, I just
cleaned up my prototype a bit.

The main complexity here is that APIFlask does not support this, so we
can't use APIFlask's decorator, and have to fallback to doing it the
flask way and then manually tell APIFlask what auth we support in the
docs (this only affects OpenAPI letting you choose the auth, it doesn't
affect anything about the actual behavior). Mildly annoying, but it's a
two-line difference from what we were doing before for any endpoint we
want this supported on.

## Validation steps
None of the endpoints actually have this enabled yet, but in addition to
the unit tests that verify the behavior, I temporarily modified the GET
opportunity endpoint to test this, just follow the details I added in
the documentation to temporarily switch it over if you want to validate
it.