[Issue #4806] Cleanup API logs for consistency (#4965)
## Summary

Fixes #4806

## Changes proposed
Adjust the API logs to have a more consistent naming structure for
common fields (ie. prefer `opportunity_id` instead of
`opportunity.opportunity.opportunity_id` or `opportunity.id`)

Adjust the JSON logger to output the message field with a different name
to work around a New Relic issue with renaming our fields

## Context for reviewers
Main motivator is so we can easily query for something like an
opportunity across all of our logs. For example, if we wanted to see
everything that interacted with a particular opportunity whether that's
in the API, a backend script, or whatever, it's all the same field name
now.