[no ticket] Fix application networking (#5037)
## Context

This is related to the NOFOs stuff, but is inside of the service module
so it would in theory impact every service. In practice, every service
already has the permission, and the NOFOs service got stood up without
it. That happened because the permission was removed here:


https://github.com/HHS/simpler-grants-gov/commit/9f0858dc064ea6dbb5ada70795f9528d971a61a3#diff-a1feeeb6afc18f0e1a2598ea4702b9d1d4810a565e24e3e696681dad91d8fab5

The ??? part is how this didn't manage to cause a problem when I was
testing that PR. Because it should have broken networking for the API
container, which was my test case. But it didn't! No idea why. But I
don't need to unpack that, the important part is that I'm fixing it now.