Add load balancer name to task definition if exists
If there is a load balancer resource, add its `dns_name` as an env
var under: `LOAD_BALANCER_DNS_NAME`

The reason we are doing this is that some (Django) apps require
a list of domains to be whitelisted as "allowed hosts", otherwise
the app will refuse to run.

So, adding this to the environment will allow us to add the load
balancer DNS name to our ALLOWED_HOSTS so that the app will boot
and not throw a 500 error.