[Issue #4695] Add form validation to the submit application endpoint (#5073)
## Summary

Fixes #4695

## Changes proposed

* Validates that the required forms on the competition exist
* Validates the json schema against the application forms answers

## Context for reviewers
Validating that required forms are filled out is pretty straightforward
- likely will be much more complex in the future (ie. user said yes to a
specific question, an optional form becomes required).

Validating that an application form is correct is also not too bad, the
main complexity addressed in this PR is communicating that information
easily to the user. Since we might have validation issues across several
forms we need a way to make it clear which form each validation issue
goes to. To solve that, the normal errors object just gets an error
object like:
```json
{
   "type": "application_form_validation",
   "message": "The application form has outstanding errors.",
   "field": "application_form_id",
   "value": "<the UUID of the application form>"
}
```
The value can be used to cross-reference it with the data object (that
up until now wasn't meaningfully used on errors). The data object will
now look like:
```json
{
   "form_validation_errors": {
       "<the UUID of the application form>": [{
                    "field": "$.name",
                    "message": "5 is not of type 'string'",
                    "type": "type",
                    "value": null,
            }]
   }
}
```
The format of the validation errors

## Validation steps
Endpoints are usable locally. First you'll need to authenticate, on the
localhost:8080/docs page, go to the login endpoint and there is a link
you can click to authenticate, grab the full token from the response
page and set it for the `ApiJwtAuth`.

Make sure you've recently run `make db-seed-local` which will create a
competition with a form, this competition always has a static ID at the
moment of `fd7f5921-9585-48a5-ab0f-e726f4d1ef94`.

Go to the start application endpoint and pass it the above competition
ID.

If you take the application ID from the start application endpoint
response and immediately try to submit it, you should get an error
roughly like:
```json
{
  "data": {},
  "errors": [
    {
      "field": "form_id",
      "message": "Form Test form is required",
      "type": "missing_required_form",
      "value": "3421aeb3-8013-449a-8a04-a1d88e1b32bc"
    },
    {
      "field": "form_id",
      "message": "Form Application For Federal Assistance SFS 424 - Individual is required",
      "type": "missing_required_form",
      "value": "e1cffadb-b316-4685-81b3-a5a43cce3c5a"
    }
  ],
  "internal_request_id": "35e2a368-1220-44fd-a34a-a1abf0864746",
  "message": "The application has issues in its form responses.",
  "status_code": 422
}
```

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>