[#5065] Add load balancer name to task definition (if load balancer exists) (#5101)
## Summary

Makes progress on #5065 

If there is a load balancer resource, add its `dns_name` as an env var
under: `LOAD_BALANCER_DNS_NAME`

## Context for reviewers

The reason we are doing this is that some (Django) apps require [a list
of domains to be whitelisted as "allowed
hosts"](https://docs.djangoproject.com/en/5.2/ref/settings/#allowed-hosts),
otherwise the app will refuse to run in production.

So, adding this to the environment will allow us to add the load
balancer DNS name to our ALLOWED_HOSTS so that the app will boot up and
not throw a 500 error.