[Issue #4985] Don't allow a user to modify an application if it's not in-progress (#5119)
## Summary

Fixes #4985

## Changes proposed

This applies to the endpoints for submitting an application, as well as
updating an application, and updating an application form.

Cleaned up some of the unit tests to use user/auth token utils instead
of redoing the logic each time (a bit more cleanup could be done, but
I'll separate that out later). Also removed some of the date freezing
logic because it broke the auth cleanup - not necessary to freeze time
if we already compare to "today".

## Context for reviewers

Once a user submits an app, they shouldn't be able to modify it. Pretty
uneventful change, we already had the logic on the submit endpoint, just
moved it to a central location and reused it.