[Issue #5080] Add more fields to the competition response schema (#5146)
## Summary

Fixes #5080 and #5131

## Changes proposed

Adds additional fields to the competition response schema, most notably
an `is_open` flag

Refactor some validation logic around how we determine if a competition
is open to be more centralized on the DB model itself

## Context for reviewers

Most of the fields come straight from the competition and are
uneventful. The `is_open` flag is a calculated field that checks the
competition open + close dates (and optionally grace period) to see if a
competition is open. We already had implemented validations to check
this field to block users from starting/submitting an application, but
now the field is calculated in one central location that those
validations now check instead. This required slightly rewording the
error message, but its behavior is the same.

## Validation steps
Running `make db-seed-local` will create a competition with a static ID
of `fd7f5921-9585-48a5-ab0f-e726f4d1ef94` that you can query.

The GET /competition endpoint returns something like:
```json
{
  "data": {
    "closing_date": "2025-06-07",
    "competition_forms": [...], // trimmed for length
    "competition_id": "fd7f5921-9585-48a5-ab0f-e726f4d1ef94",
    "competition_title": "Charge admit hard.",
    "contact_info": null,
    "is_open": true,
    "open_to_applicants": [
      "organization",
      "individual"
    ],
    "opening_date": "2025-05-08",
    "opportunity_assistance_listing": {
      "assistance_listing_number": "14.337",
      "program_title": "Leonard, Lowe and Vang"
    },
    "opportunity_id": 112
  },
  "message": "Success",
  "status_code": 200
}
```

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>