[#5141] Ignore warning about ":latest" tags in NOFO docker images (#5156)
## Summary

Fixes #5141 

This is the last issue raised by the vuln scanners, but for this one we
do want to keep the `:latest` tag.

Also, this is a `warning` level message, not a `FATAL` message.

## Changes proposed

Create a temporary .dockleignore file so that we can pass in this one
rule.

## Context for reviewers

The `:latest` tag is part of our CI/CD workflow, so removing it would
not be good for us.