[Issue #4862] Add a button that resets the feature flags cookie (#5158)
## Summary

Fixes #4862 

Builds on prior work by Austin

## Changes proposed

Adds a button on the Feature Flag page that uses an a tag link to fully
refresh the app while reseting the Feature Flags back to defaults.

## Context for reviewers

Do a vanilla a tag so that we get the full app reload after reseting the
feature flags cookie, since the flags can adjust nav, etc.

## Validation steps

1. Checkout the branch
2. Run using `npm run dev`
3. Go to http://localhost:3000/dev/feature-flags
4. Try enabling and disabling things to get out of sync with the current
default settings
5. Use the Reset button to verify they go back to defaults.