[Unticketed] Fix a test broken by another PR (#5193)
## Summary


## Changes proposed

Fixes a test that broke when two PRs got merged to main together

## Context for reviewers
Changed some types in one PR, and used the old types in another, broke
when they merged together. Even had a note on the test that this would
happen, and forgot to fix it. Simple enough to fix.