[Issue #4986] Add an application form status to the response model for relevant endpoints (#5201)
## Summary

Fixes #4986

## Changes proposed
Calculate an applicant form status and return it in:
* GET /applications/:appId
* PUT /applications/:appId/form/:formId
* GET /applications/:appId/application_forms/:appFormId

## Context for reviewers
There is still a bit of ongoing discussion with product/design about the
exact way we calculate the form status, this might shift a bit, but
that's not a blocker for getting something out.

The statuses are meant to help guide the user in filling out forms, even
if every form has an application status of "complete" the app could
still be rejected on submission for validations outside of the forms
(eg. the submission window has closed).

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>