[Issue #5078] Create an endpoint for adding an attachment to an application (#5191)
## Summary

Fixes #5078

## Changes proposed
Create `POST /applications/:application_id/attachments` which allows
adding a file to your application.

## Context for reviewers
This is our first endpoint for uploading files, so a few utilities
needed to be setup to support that. A few things to call out:
* To support uploading large files, we don't want to use a standard JSON
body request, but instead use a "form" request which allows for
multi-part upload of the file which should perform better than having to
base64 encode a file into a JSON object.
* The file we upload to s3 has a slightly modified filename to keep
things simple (removes spaces and non-ascii characters) - the users
actual file name is stored in the DB so on the UI it matches whatever
they might have on disk.
* The mimetype of the file is currently pulled from the form and is set
by the browser actually, it can be a generic "application/octet-stream"
if not specified which is fine. We could look into guessing it, but I
started down that path and it's definitely its own dedicated ticket to
setup.

## Validation steps
If you create an application locally, you can easily add files to it via
the swagger docs - it handles the form type just fine and you can select
files:
<img width="749" alt="Screenshot 2025-05-29 at 10 50 58 AM"
src="https://github.com/user-attachments/assets/6f559e7f-b142-4108-94a1-f904c88b1173"
/>

I did also do some testing of running curls to upload files (swagger
will give you the curl command) and removed the content type - it still
works fine but just makes things be "application/octet-stream" in most
cases.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>