[Issue #5135] In the create application endpoint, initialize the application form objects (#5221)
## Summary

Fixes #5135

## Changes proposed
When we create a new application, create an empty application form
record for each competition form

Restructure some validation logic that assumed application forms would
be missing in normal circumstances

## Context for reviewers
We want to pre-initialize the application forms so a user like our
frontend fetching an application gets the app forms with all forms on
the competition automatically. This is purely to make the data more
obvious to a user / not have things be missing because a user hasn't
started an application form.

This does mean that some validation elsewhere around required forms had
to change a bit as a non-required form will always exist, and I had to
adjust the validation logic accordingly. To make things easier, we want
to be able to iterate over the app forms, and to verify something isn't
missing I added a new validation checker for an app form having not been
created (shouldn't happen in normal usage, but as a safety net).

## Validation steps
`make db-seed-local` will create a competition with ID
`fd7f5921-9585-48a5-ab0f-e726f4d1ef94`. If you start an application for
that competition, it will always make an application form for each
competition now. Both competition forms get returned, and a warning
exists for both of them for not being started while being required.

```json
{
  "data": {
    "application_forms": [
      {
        "application_form_id": "bf6e10b2-2e3a-4281-a800-8c765bb61462",
        "application_form_status": "not_started",
        "application_id": "8a4c2a30-411a-41f2-8655-a9e67a085e63",
        "application_response": {},
        "form_id": "2298bb0f-0656-479c-8725-c17e4f435f51"
      },
      {
        "application_form_id": "825f2628-9e58-4523-8ce1-2e6802e0e14f",
        "application_form_status": "not_started",
        "application_id": "8a4c2a30-411a-41f2-8655-a9e67a085e63",
        "application_response": {},
        "form_id": "abd642a0-b3da-4e18-9e9d-57babd4614d7"
      }
    ]
   // a bunch I'm trimming for brevity
  }
  "warnings": [
    {
      "field": "form_id",
      "message": "Form Test form is required",
      "type": "missing_required_form",
      "value": "2298bb0f-0656-479c-8725-c17e4f435f51"
    },
    {
      "field": "form_id",
      "message": "Form Application For Federal Assistance SFS 424 - Individual is required",
      "type": "missing_required_form",
      "value": "abd642a0-b3da-4e18-9e9d-57babd4614d7"
    }
  ]
}
```

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>